refactor(oxfmt): Arrange cli/js_config dir (#21078)
- Add `/js_config/` dir
- Use inline tests
- Always return `Error` instance
