+++ apps/oxfmt/src-js/cli.ts
@@ -7,7 +7,7 @@ import {
   formatEmbeddedDoc,
   sortTailwindClasses,
 } from "./cli/worker-proxy";
-import { loadJsConfig } from "./cli/js_config";
+import { loadJsConfig } from "./cli/js_config/index";
 
 // napi-JS `oxfmt` CLI entry point
 // See also `run_cli()` function in `./src/main_napi.rs`
