+++ apps/oxfmt/src-js/cli/js_config/index.ts
@@ -1,6 +1,6 @@
 import { basename as pathBasename } from "node:path";
 import { pathToFileURL } from "node:url";
-import { getUnsupportedTypeScriptModuleLoadHintForError } from "./node_version";
+import { getUnsupportedTypeScriptModuleLoadHint } from "./node_version";
 
 const isObject = (v: unknown) => typeof v === "object" && v !== null && !Array.isArray(v);
 
@@ -24,13 +24,11 @@ export async function loadJsConfig(path: string): Promise<object | null> {
   const fileUrl = pathToFileURL(path);
   fileUrl.searchParams.set("cache", Date.now().toString());
 
-  let config: unknown;
-  try {
-    ({ default: config } = await import(fileUrl.href));
-  } catch (err) {
-    const unsupportedNodeHint = getUnsupportedTypeScriptModuleLoadHintForError(err, path);
-    throw unsupportedNodeHint ?? err;
-  }
+  const { default: config } = await import(fileUrl.href).catch((err) => {
+    const hint = getUnsupportedTypeScriptModuleLoadHint(err, path);
+    if (hint && err instanceof Error) err.message += `\n\n${hint}`;
+    throw err;
+  });
 
   if (config === undefined) throw new Error("Configuration file has no default export.");
