--- apps/oxfmt/test/cli/js_config/node_version.test.ts
@@ -1,41 +1,0 @@
-import { describe, expect, it } from "vitest";
-import {
-  NODE_TYPESCRIPT_SUPPORT_RANGE,
-  getUnsupportedTypeScriptModuleLoadHintForError,
-  isTypeScriptModuleSpecifier,
-} from "../../../src-js/cli/node_version";
-
-describe("node_version", () => {
-  it("detects supported TypeScript config specifiers", () => {
-    expect(isTypeScriptModuleSpecifier("/tmp/oxfmt.config.ts")).toBe(true);
-    expect(isTypeScriptModuleSpecifier("/tmp/oxfmt.config.mts")).toBe(true);
-    expect(isTypeScriptModuleSpecifier("/tmp/oxfmt.config.cts")).toBe(true);
-    expect(isTypeScriptModuleSpecifier("file:///tmp/oxfmt.config.ts")).toBe(true);
-    expect(isTypeScriptModuleSpecifier("/tmp/oxfmt.config.js")).toBe(false);
-  });
-
-  it("adds a node version hint for unsupported TypeScript module loading", () => {
-    const err = new TypeError(
-      'Unknown file extension ".ts" for /tmp/oxfmt.config.ts',
-    ) as TypeError & {
-      code?: string;
-    };
-    err.code = "ERR_UNKNOWN_FILE_EXTENSION";
-
-    expect(
-      getUnsupportedTypeScriptModuleLoadHintForError(err, "/tmp/oxfmt.config.ts", "v22.11.0"),
-    ).toBe(
-      `TypeError: Unknown file extension ".ts" for /tmp/oxfmt.config.ts\n\nTypeScript config files require Node.js ${NODE_TYPESCRIPT_SUPPORT_RANGE}.\nDetected Node.js v22.11.0.\nPlease upgrade Node.js or use a JSON config file instead.`,
-    );
-  });
-
-  it("does not add the hint for non-TypeScript specifiers or unrelated errors", () => {
-    const err = new Error("Cannot find package");
-    expect(getUnsupportedTypeScriptModuleLoadHintForError(err, "/tmp/oxfmt.config.ts")).toBeNull();
-
-    const unknownExtension = new TypeError('Unknown file extension ".ts"');
-    expect(
-      getUnsupportedTypeScriptModuleLoadHintForError(unknownExtension, "/tmp/oxfmt.config.js"),
-    ).toBeNull();
-  });
-});
