+++ apps/oxfmt/vitest.config.ts
@@ -1,8 +1,9 @@
-import { configDefaults, defineConfig } from "vitest/config";
+import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    exclude: [...configDefaults.exclude],
+    include: ["./test/**/*.test.ts"],
+    includeSource: ["./src-js/**/*.ts"],
     snapshotFormat: {
       escapeString: false,
       printBasicPrototype: false,
